A-CheckStyle: Update code to comply with codestyle
