+++ build.gradle
@@ -1,6 +1,7 @@
 plugins {
     id 'java'
     id 'application'
+    id 'checkstyle'
     id 'com.github.johnrengelman.shadow' version '5.1.0'
 }
 
@@ -52,4 +53,8 @@ shadowJar {
 
 run{
     standardInput = System.in
+}
+
+checkstyle{
+    toolVersion = '10.2'
 }
