--- src/main/java/dobby/UI/DialogBox.java
@@ -1,88 +1,0 @@
-package dobby.UI;
-
-import java.io.IOException;
-import java.util.Collections;
-
-import javafx.collections.FXCollections;
-import javafx.collections.ObservableList;
-import javafx.fxml.FXML;
-import javafx.fxml.FXMLLoader;
-import javafx.geometry.Pos;
-import javafx.scene.Node;
-import javafx.scene.control.Label;
-import javafx.scene.image.Image;
-import javafx.scene.image.ImageView;
-import javafx.scene.layout.HBox;
-import javafx.scene.text.TextAlignment;
-import javafx.scene.text.TextFlow;
-
-/**
- * This control represents a dialog box consisting of an ImageView to represent the speaker's face and a label
- * containing text from the speaker.
- * <p>
- * Adapted from <a href="https://se-education.org/guides/tutorials/javaFxPart4.html">se-education</a>
- */
-public class DialogBox extends HBox {
-    @FXML
-    private Label dialog;
-    @FXML
-    private ImageView displayPicture;
-    @FXML
-    private TextFlow textFlow;
-
-    /**
-     * Creates a dialog box with the given text and image.
-     * @param text the text to be displayed in the dialog box
-     * @param img the image to be displayed in the dialog box
-     */
-    private DialogBox(String text, Image img) {
-        try {
-            FXMLLoader fxmlLoader = new FXMLLoader(MainWindow.class.getResource("/view/DialogBox.fxml"));
-            fxmlLoader.setController(this);
-            fxmlLoader.setRoot(this);
-            fxmlLoader.load();
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-
-        dialog.setText(text);
-        displayPicture.setImage(img);
-    }
-
-    /**
-     * Flips the dialog box such that the ImageView is on the left and text on the right.
-     */
-    private void flip() {
-        ObservableList<Node> tmp = FXCollections.observableArrayList(this.getChildren());
-        Collections.reverse(tmp);
-        getChildren().setAll(tmp);
-        setAlignment(Pos.TOP_LEFT);
-        if (textFlow.getTextAlignment() == TextAlignment.RIGHT) {
-            textFlow.setTextAlignment(TextAlignment.LEFT);
-        } else {
-            textFlow.setTextAlignment(TextAlignment.RIGHT);
-        }
-    }
-
-    /**
-     * Creates a dialog box for the user.
-     * @param text the text to be displayed
-     * @param img the image to be displayed
-     * @return a dialog box for the user
-     */
-    public static DialogBox getUserDialog(String text, Image img) {
-        return new DialogBox(text, img);
-    }
-
-    /**
-     * Creates a dialog box for Duke.
-     * @param text the text to be displayed
-     * @param img the image to be displayed
-     * @return a dialog box for Duke
-     */
-    public static DialogBox getDukeDialog(String text, Image img) {
-        var db = new DialogBox(text, img);
-        db.flip();
-        return db;
-    }
-}
