+++ src/main/java/dobby/ui/Launcher.java
@@ -1,15 +1,15 @@
-package dobby.UI;
+package dobby.ui;
 
 import dobby.Main;
 import javafx.application.Application;
 
 /**
- * Launches Duke
+ * Launches Duke.
  * Adapted from <a href="https://se-education.org/guides/tutorials/javaFxPart4.html">se-education</a>
  */
 public class Launcher {
-    public static void main(String[] args) {
-        Application.launch(Main.class, args);
-        System.exit(0);
-    }
+  public static void main(String[] args) {
+    Application.launch(Main.class, args);
+    System.exit(0);
+  }
 }
