+++ src/main/resources/view/MainWindow.fxml
@@ -3,14 +3,20 @@
 <?import javafx.scene.control.*?>
 <?import javafx.scene.layout.*?>
 
-<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="600.0" prefWidth="701.0" xmlns="http://javafx.com/javafx/17.0.2-ea" xmlns:fx="http://javafx.com/fxml/1" fx:controller="dobby.UI.MainWindow">
+<AnchorPane maxHeight="-Infinity" maxWidth="-Infinity" minHeight="-Infinity" minWidth="-Infinity" prefHeight="600.0"
+            prefWidth="701.0" xmlns="http://javafx.com/javafx/17.0.2-ea" xmlns:fx="http://javafx.com/fxml/1"
+            fx:controller="dobby.ui.MainWindow">
     <children>
-        <TextField fx:id="userInput" layoutY="558.0" onAction="#handleUserInput" prefHeight="41.0" prefWidth="548.0" AnchorPane.bottomAnchor="1.0" />
-        <Button fx:id="sendButton" layoutX="624.0" layoutY="557.0" mnemonicParsing="false" onAction="#handleUserInput" prefHeight="41.0" prefWidth="76.0" text="Send" />
-        <Button fx:id="helpButton" layoutX="548.0" layoutY="557.0" mnemonicParsing="false" onAction="#printHelp" prefHeight="41.0" prefWidth="76.0" text="Help" />
-        <ScrollPane fx:id="scrollPane" hbarPolicy="NEVER" hvalue="1.0" prefHeight="557.0" prefWidth="700.0" vvalue="1.0">
+        <TextField fx:id="userInput" layoutY="558.0" onAction="#handleUserInput" prefHeight="41.0" prefWidth="548.0"
+                   AnchorPane.bottomAnchor="1.0"/>
+        <Button fx:id="sendButton" layoutX="624.0" layoutY="557.0" mnemonicParsing="false" onAction="#handleUserInput"
+                prefHeight="41.0" prefWidth="76.0" text="Send"/>
+        <Button fx:id="helpButton" layoutX="548.0" layoutY="557.0" mnemonicParsing="false" onAction="#printHelp"
+                prefHeight="41.0" prefWidth="76.0" text="Help"/>
+        <ScrollPane fx:id="scrollPane" hbarPolicy="NEVER" hvalue="1.0" prefHeight="557.0" prefWidth="700.0"
+                    vvalue="1.0">
             <content>
-                <VBox fx:id="dialogContainer" prefHeight="552.0" prefWidth="702.0" />
+                <VBox fx:id="dialogContainer" prefHeight="552.0" prefWidth="702.0"/>
             </content>
         </ScrollPane>
     </children>
